perf improvement in sorted-numbers?
sorted-numbers? is called once at the beginning of the permutations
function in order to determine which version of the permutations
algorithm to use -- a more efficient algorithm can be chosen if the
numbers are already sorted.  Since this function is only called once,
improving its performance will have a negligible effect on the overall
performance of permutations.  But this is an improvement, so why not?